requirements_mac.txt
astroid==2.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
isort==4.3.21
iterfzf==0.5.0.20.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.18.4
pandas==1.1.3
Pillow==8.0.0
plotly==4.8.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pylint==2.5.2
pyparsing==2.4.7
pysmb==1.2.1
python-dateutil==2.8.1
pytz==2020.1
qprompt==0.16.0
requests==2.23.0
retrying==1.3.3
scipy==1.5.3
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
